+++ inifile.c
@@ -86,7 +86,7 @@ ini_section *read_ini(char *ifile_name) {
         } while((*d++ = *s++));
 
         /* Get section */
-        if (sscanf(buffer, "[%[a-zA-Z0-9_\t +()-]]", section) == 1) {
+        if (sscanf(buffer, "[%255[a-zA-Z0-9_\t +()-]]", section) == 1) {
             printl(LOG_VERB, "LN: [%d] S: [%s]", ln, section);
 
             /* Current section to use */